update epics to use graphql
