+++ frontend/src/actions/userActions.js
@@ -1,4 +1,4 @@
-import { LOGIN_USER, LOGGED_IN_USER, LOGOUT_USER, REGISTER_USER } from '../constants/UserActionTypes'
+import { LOGIN_USER, LOGGED_IN_USER, LOGOUT_USER, REGISTER_USER, GET_USER_STATUS, USER_STATUS_RECEIVED } from '../constants/UserActionTypes'
 
 export function registerUser (credentials) {
   return {
@@ -24,5 +24,19 @@ export function loggedInUser (userObj) {
 export function logoutUser () {
   return {
     type: LOGOUT_USER
+  }
+}
+
+export function getUserStatus (userDetails) {
+  return {
+    type: GET_USER_STATUS,
+    payload: userDetails
+  }
+}
+
+export function userStatusReceived (responseDetails) {
+  return {
+    type: USER_STATUS_RECEIVED,
+    responseDetails
   }
 }
