+++ frontend/src/constants/NetworkActionTypes.js
@@ -1,4 +1,2 @@
 export const GET_NETWORK_STATUS = 'GET_NETWORK_STATUS'
 export const NETWORK_STATUS_RECEIVED = 'NETWORK_STATUS_RECEIVED'
-export const GET_USER_STATUS = 'GET_USER_STATUS'
-export const USER_STATUS_RECEIVED = 'USER_STATUS_RECEIVED'
