+++ frontend/src/constants/UserActionTypes.js
@@ -2,3 +2,5 @@ export const REGISTER_USER = 'REGISTER_USER'
 export const LOGIN_USER = 'LOGIN_USER'
 export const LOGGED_IN_USER = 'LOGGED_IN_USER'
 export const LOGOUT_USER = 'LOGOUT_USER'
+export const GET_USER_STATUS = 'GET_USER_STATUS'
+export const USER_STATUS_RECEIVED = 'USER_STATUS_RECEIVED'
