+++ frontend/src/epics/project.js
@@ -1,25 +1,11 @@
 import { GET_PROPOSED_PROJECTS } from '../constants/ProjectActionTypes'
 import { proposedProjectsReceived } from '../actions/projectActions'
-
-import { combineEpics } from 'redux-observable'
-import { Observable } from 'rxjs'
 import { map, mergeMap } from 'rxjs/operators'
-
-let getConfig = {
-  method: 'GET',
-  headers: new Headers(),
-  mode: 'cors',
-  cache: 'default'
-}
-
-export async function fetchService (url) {
-  let response = await fetch(url, getConfig)
-  return response.json()
-}
+import { client } from '../index'
+import gql from 'graphql-tag'
 
 export const getProposedProjectsEpic = action$ =>
   action$.ofType(GET_PROPOSED_PROJECTS).pipe(
-    mergeMap(action => Observable.from(fetchService(`/api/projects/all?state=1`))),
-    map(res => Observable.of(res)),
+    mergeMap(action => client.query({query: gql`{ user(state: 1){}}`})),
     map(result => proposedProjectsReceived(result))
   )
